Renames and split eact type to its own class
