+++ sample/Server/Startup.cs
@@ -1,4 +1,4 @@
-﻿using GRPCServer.Dotnet;
+﻿using Grpc.AspNetCore;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
