+++ src/Grpc.AspNetCore/Grpc.AspNetCore.csproj
@@ -5,12 +5,12 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Grpc.Common\Grpc.Common.csproj" />
+    <PackageReference Include="Microsoft.AspNetCore.App" Version="3.0.0-preview-18607-0108" />
+    <PackageReference Include="Google.Protobuf" Version="3.6.1" />
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App" Version="3.0.0-preview-18607-0108" />
-    <PackageReference Include="Google.Protobuf" Version="3.6.1" />
+    <ProjectReference Include="..\Grpc.Common\Grpc.Core.Api.csproj" />
   </ItemGroup>
 
 </Project>
