+++ src/Grpc.AspNetCore/IServerCallHandler.cs
@@ -1,0 +1,11 @@
+﻿
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
+
+namespace Grpc.AspNetCore
+{
+    interface IServerCallHandler
+    {
+        Task HandleCallAsync(HttpContext httpContext);
+    }
+}
